requirements.txt
# ============================================================
# JARVIS - Full Requirements
# Install with: pip install -r requirements.txt
# ============================================================

# GUI
PySide6>=6.6.0

# Voice Input / Output
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13

# Spell Correction / NLP
textblob>=0.18.0

# LLM (Ollama local models)
ollama>=0.1.7

# Memory / Database
# SQLite is built into Python — no install needed

# HTTP Requests (APIs)
requests>=2.31.0

# System Monitoring
psutil>=5.9.0

# Automation
pyautogui>=0.9.54

# Environment Variables
python-dotenv>=1.0.0

# Future: Semantic memory
# chromadb>=0.4.0

# Future: Web Search
# duckduckgo-search>=3.9.6
